Added key and mouse codes, input polling and basic imgui layer, will be changed next

+++ Leaf/src/Leaf/Events/ApplicationEvent.h
@@ -16,8 +16,8 @@ namespace Leaf {
 			return s.str();
 		}
 
-		inline float GetHeight() const { return m_Height; }
-		inline float GetWidth() const { return m_Width; }
+		inline uint32_t GetHeight() const { return m_Height; }
+		inline uint32_t GetWidth() const { return m_Width; }
 
 		SET_EVENT_TYPE(OnWindowResize)
 		SET_EVENT_CATEGORY(ApplicationEvent)
